Refactoring, removing unused code
